Refactor: 타입 정리
