--- src/types/analysis.types.ts
@@ -1,5 +1,0 @@
-export interface Analysis {
-    analysisId: number;
-    answer: string;
-    video: string;
-  }
