+++ src/types/api.familyedit.types.ts
@@ -1,0 +1,6 @@
+// /parent/mypage/edit
+// /parent/mypage/edit/familyName
+// /parent/mypage/edit/name
+// /parent/mypage/edit/add
+// /parent/mypage/edit/{childId}
+// /parent/mypage/edit/delete
