+++ src/types/api.notice.types.ts
@@ -1,0 +1,7 @@
+// /parent/notice
+// /parent/notice/readAll
+// /parent/notice/accept
+// /parent/notice/decline
+// /parent/{childId}/pressNotice
+
+
