+++ src/types/api.parententer.types.ts
@@ -1,0 +1,27 @@
+// /parent/mypage
+// /parent/{childId}/banner
+// /parent/{childId}/stat/cumulative
+// /parent/{childId}/keywords
+
+
+// /parent/{childId}/stat/subject -키워드 누르면
+
+
+import { Category } from './stats.types';
+
+/**
+ * /parent/{childId}/keywords API의 응답 데이터 타입입니다.
+ * (카테고리 정보를 답변 개수 순위에 맞게 반환)
+ *
+ * @example
+ * {
+ * keywords: [
+ * { category: "가족", count: 10, positiveScore: 80 },
+ * { category: "학교", count: 5, positiveScore: 40 }
+ * ],
+ * }
+ */
+
+export interface KeywordResponse {
+  keywords: Category[];
+}
