+++ src/types/api.stats.types.ts
@@ -1,0 +1,11 @@
+// 답변 기록
+// /parent/{childId}/home subjects
+// 예정 질문
+// /parent/{childId}/scheduled
+// /parent/{childId}/questions/edit
+// /parent/{childId}/questions/reroll
+// 달력
+// /parent/{childId}/stat/emotion
+// /parent/{childId}/subjects 
+// 상세질문 리스트
+// /parent/{childId}/{subjectId}
