+++ src/types/index.ts
@@ -1,6 +1,12 @@
-export * from './analysis.types';
-export * from './child.types';
-export * from './question.types';
+
+//부모페이지 api Respose 타입들 
+export * from './api.start.types';
+export * from './api.stats.types';
+export * from './api.parententer.types';
+export * from './api.familyedit.types';
+export * from './api.notice.types';
+
+//기타 공통 타입들
+export * from './people.types';
 export * from './subject.types';
-export * from './parent.types';
-export * from './api.types';
+export * from './stats.types';
