--- src/types/parent.types.ts
@@ -1,6 +1,0 @@
-export interface Parent {
-  id: number;
-  name: string;
-  email?: string;
-  own?:boolean;
-}
