--- src/types/question.types.ts
@@ -1,7 +1,0 @@
-export interface Question {
-  questionId: number;
-  text: string;
-  video: string;
-  image: string;
-  answer: string;
-}
