+++ src/types/stats.types.ts
@@ -1,0 +1,5 @@
+export interface Category {
+  category: string;
+  count: number;
+  positiveScore: number;
+}
